+++ js_modules/dagster-ui/packages/ui-core/client.json
@@ -10,14 +10,14 @@
   "SidebarAssetQuery": "5f877e6818194f30b948bda42fea8e5ef954461ce44bb45d7d201a677ec30487",
   "AssetGraphQuery": "26030b5c565bdc4d84b54b2c9a7e8172562cf7434912511768bde20875d47b44",
   "AssetForNavigationQuery": "eb695ab88044ddd7068ea0dc1e2482eaba1fcb11b83de11050ff52f55e83ed3d",
+  "AssetCatalogTableQuery": "5422b3da3382bc9ac3122990d11110bc63f8b27b4389e4411b95b0bb131f987b",
+  "AssetCatalogGroupTableQuery": "77c8aa916469e8e1af76ae8d31d8feb512b743e2996180111368e0e5fe17cc90",
   "AssetDefinitionCollisionQuery": "84027ea05480797a69eb150ce07ac7dfd40d6007a8107f90a6c558cf3a2662f5",
   "AssetGroupMetadataQuery": "649fd8034ea453acebec574365ec19b678930cdbc976f5da2addfe40f985886f",
   "AssetMaterializationUpstreamQuery": "754bab88738acc8d310c71f577ac3cf06dc57950bb1f98a18844e6e00bae756d",
   "AssetPartitionDetailQuery": "a5a3c782894d7365cbf791c57bdec279cbccd809cd2d0dc5fd31df1e6237b927",
   "AssetPartitionStaleQuery": "4215f4014e9d7592142e1775c4b07377703e913389396f9ca14dc6bb779ce764",
   "AssetViewDefinitionQuery": "6c820a3851557b1cdf92373901107c6d29ea94f9da70e111b8ebe47e561da58b",
-  "AssetCatalogTableQuery": "5422b3da3382bc9ac3122990d11110bc63f8b27b4389e4411b95b0bb131f987b",
-  "AssetCatalogGroupTableQuery": "77c8aa916469e8e1af76ae8d31d8feb512b743e2996180111368e0e5fe17cc90",
   "AssetsOverviewRootQuery": "77ab0417c979b92c9ec01cd76a0f49b59f5b8ce7af775cab7e9b3e57b7871f7d",
   "RunStatusAndPartitionKey": "4642abda7da52fb70cc0a47c19cd5bf2fd8b854bb104b6a73eb8545fcd0146b2",
   "RunStatusOnlyQuery": "e0000c8f2600dbe29f305febb04cca005e08da6a7ce03ec20476c59d607495c0",
